Use alternate method to determine disk size
query_virtual_disk_geometry seems to not work for all versions of
vCenter

We have to look up the "*-flat.vmdk" version because the file_size
matches the actual disk capacity.

We moved most of the disk finding logic to the VSphere::Client and plan
to integration test it there.

Signed-off-by: Colin OByrne <cobyrne@pivotal.io>