Adds integration coverage for the VSphereCloud::Client disk methods
We're faking the client in unit tests, and the lifecycle_specs are more
concerned with high-level flows, so we want to add more specific
coverage for our client's integration with vSphere.

This spec depends on a subset of the environment variables used by the
lifecycle specs, so it should be easy to run in development and on ci.

Signed-off-by: Kihyeon Kim <kihyeon.kim@gmail.com>