Verify that user has limited permissions for running lifecycle.
The CPI should only need a very limited set of permissions on the
root_folder in order to work. We don't need our lifecycles to pass a
user with too many permissions so we are adding an assertion that the
user being tested only has the minimal access it should.

[#91137208](https://www.pivotaltracker.com/story/show/91137208)

Signed-off-by: Jonathan Barnes <jbarnes@pivotal.io>