Allows lifecycle task to run on either vsphere worker
Rather than having lifecycle-v5.1/6.0 run on their own respective
workers we are going to have either lifecycle run against either worker.
We are putting the tagging of the task in the task configure file so
that fly exec's against prod will work and fly exec's against local will
warn that there is no such worker allowing the developer to tweak the
task configuration.

[#105585636](https://www.pivotaltracker.com/story/show/105585636)