Removes disk provisioning types which are not supported by vSphere 5.1.
Removes disk types which manual testing has shown to not work in the CI
environment. Adds no lifecycle tests as the only way to verify the disk
type would be through the vSphere client as disk type is not exposed
outside of the client. Updates unit tests to reflect supported types.

[#105338800](https://www.pivotaltracker.com/story/show/105338800)

Signed-off-by: Beyhan Veli <beyhan.veli@sap.com>