vSphere CPI now looks in entire datacenter for VMs and stemcells
Previously, the vSphere CPI searched for VMs and stemcells using an
inventory path that involved a VM or template folder name in the vm
folder in the datacenter. This inventory path changed depending on
whether or not the vSphere resource pool is specified in the clusters
section of the global CPI configuration, so changing the CPI
configuration caused the CPI to be unable to find the VMs or stemcells
of deployments.

The vSphere CPI now looks for any VM or stemcell with a specified name
in the entire datacenter by getting a list of all VMs in the
datacenter and finding the VM with a certain name.

[#106999568](https://www.pivotaltracker.com/story/show/106999568)

Signed-off-by: Edwin Xie <exie@pivotal.io>