'get' steps for lifecycle-5.1 use the vsphere-v5.1 worker
In order to take advantage of baggageclaim resource caching, we need the
resource-getting workers to be the same as the task workers. There are a
couple of things we want to test: (1) see if resource fetching on the
same worker as the task running will improve job execution times and (2)
vSphere is currently the only worker we have enabled resource caching on
so we want to take advantage of it.

[#108131232](https://www.pivotaltracker.com/story/show/108131232)

Signed-off-by: Aaron Hurley <ahurley@pivotal.io>