Deleting a VM will rename persistent disks if they were renamed outside of BOSH
Currently, the CPI has no way to find a persistent if it is renamed
outside of BOSH (such as a storage vMotion). This commit introduces a
vApp property tag that is attached to VMs when a persistent disk is
attached. Upon deletion of a VM, the persistent disks will get renamed
back to their original names if they had been renamed prior to the last
bosh deploy.

Client#move_disk uses virtual_disk#move_virtual_disk instead of
file_manager#move_file as this seems like a better way to move (and
rename) virtual disks.

Two new permissions were added to the role that runs lifecycle tests
because adding vApp properties require these.

[#109550896](https://www.pivotaltracker.com/story/show/109550896)

Signed-off-by: Thinh Nguyen <thinh.nguyen@emc.com>