Correct expected error in host_maintenance_mode_spec
When a host enters the maintenace mode, the respective cluster's runtime info
should not account for that host's memory. Due to some race condition in
which the "runtime info is getting updated later and is queried faster"
causes in faulty value being fetched and differing errors to get raised.

This change just accounts for expected error string in integration test. The
code will always raise an exception , just the point of discovery of no
valid placement is ambigous.