CI: Pipelines test NSX-T Policy API
This introduces the Policy API tests. We test the Policy API via BATS,
which exercises the network portion among other things. We deploy to an
NSX-T segment (a.k.a. vSphere portgroup, a.k.a. logical switch).

We have a separate Concourse tasks to deploy with Policy; it deploys to
the NSX-T segment instead of "VM Network". It uses the "30.0/16" instead
of the "192.168/16" address space. It deploys the BOSH Director with the
necessary NSX-T properties (user, password, cert, `use_policy_api`).

We modified `main.yml.erb` (Concourse pipeline source) to use the
double-parentheses syntax `(( ...))` rather than the double-curly-brace
syntax `{{ ... }}`. The reasons are twofold: the double-curly-brace
syntax is deprecated, but, more importantly, Concourse only interpolates
from Vault if we use the double-parentheses, and we plan to use Vault to
store our secrets.

[#176769026](https://www.pivotaltracker.com/story/show/176769026)

Co-authored-by: Michael Oleske <moleske@vmware.com>
Co-authored-by: Brian Cunnie <bcunnie@vmware.com>