Use host group memory for cluster preference
Cluster free memory is used for filtering the clusters. We split the
memory into two types: cluster and host group free memory. When the
cluster has host groups specified the policy will define which memory is
used for filtering out available clusters for required memory and which
memory is used for preference of the cluster.

In case of host group rule "MUST" host group must have the required
memory and clusters that don't have host groups with the required memory
won't be used.

In case of host group rule "SHOULD" host group might not have the
required memory. We pick the cluster if it has the required memory in
all host groups. And if specified host group doesn't have the required
memory the VM will be put on another host group on the same cluster
which has enough memory. So first cluster free memory is used to filter
out clusters and then host group memory is used to pick the host group.

[#177099523](https://www.pivotaltracker.com/story/show/177099523)

Co-authored-by: Michael Oleske <moleske@pivotal.io>
Co-authored-by: Maria Shaldybin <mariash@vmware.com>