Remove unnecessary Concourse job "delivery"
The main BOSH vSphere CPI pipeline has a job, `delivery`, whose purpose
is to mark Pivotal Tracker stories as "delivered".

As new stewards of the BOSH vSphere CPI, we feel this job is unnecessary
for us; however, we may revisit this decision should the tempo of
changes to the CPI increase.

Co-authored-by: Brian Cunnie <bcunnie@vmware.com>
Co-authored-by: Michael Oleske <moleske@vmware.com>