Put back remote_auth for VMware Identity Manager
- when adding support for [ and ], we regenerated both the manager and
policy client since this was a change that happened in a template.
However the manager client was manually changed to support VMware
Identity Manager. This went unnoticed until a bosh director deploy that
had it set caught it.  This commit properly enables a test to catch the
change.  A future story will need to resolve the fact that generating
the client will wipe this out

The policy client does not care about this value for now, so not making
a change there until we need to

[#177272017]

Authored-by: Michael Oleske <moleske@pivotal.io>