CPI `set_vm_metadata` also sets NSX-T port tags
When the
[`set_vm_metadata`](https://bosh.io/docs/cpi-api-v1-method/set-vm-metadata/)
is invoked, the metadata is set as a tag/scope on the NSX-T port of the
associated VM.

Previously the metadata was set on the vSphere vCenter, as a Custom
Attribute, and a subset of the metadata, `id`, was set on the NSX-T
port. Now all the metadata is set on the NSX-T port.

The metadata is a key-value pair. The BOSH Director has seven default
keys: `deployment`, `id`, `job`, `instance_group`, `index`, `name`, and
`created_at`. Users can set additional VM metadata, e.g. `region`.

When CPI sets the NSX-T tag, it prepends `bosh/` (e.g.
`bosh/deployment`) to the tag, and the tag's scope is set to the value
of the metadata (e.g. `cf`). This conforms with the earlier behavior for
the `id` metadata. Similarly, the tag is set on the VM's port, not on
the VM itself, which also conforms to the earlier behavior.

[#177936291] Duplicate the Ops Manager tags field as NSG tags

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>