CI: lifecycle tests: sshuttle+squid → OpenVPN
We switch from using a combination of sshuttle+squid to OpenVPN. We hope
this will accomplish two things:

1. Make the lifecycle tests more resilient
1. Speed up our lifecycle tests

The lifecycle tests have been failing at a prodigious rate since we
switched to `sshuttle` as a transport mechanism to the vSphere
environment. As an extreme example, for the job
[lifecycle-7.0-nsxt30](https://runway-ci.eng.vmware.com/teams/dap-tas-anycloud/pipelines/vsphere-cpi/jobs/lifecycle-7.0-nsxt30),
runs 66 through 75 experienced only a 10% success rate.

Also, the tests have been taking longer. For example, run 69 took 175
minutes. After switch to OpenVPN, a lifecycle run took 109 minutes: a
60% increase in speed!

We note that there are 7 lifecycle tests, and 107 workers, so the
probably that any two jobs run on the same worker is ~6%. We expect runs
that share the same worker to fail because they'll both contend for the
tunneling device & routes.

[#178271765] Fix repeated vSphere CPI CI lifecycle failures

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>