Accelerate the `info` CPI RPC call thirtyfold
We reduce the time the info CPI RPC call takes to complete. On a typical
BOSH Director, the `info` CPI RPC previously took ~10 seconds to
complete. With these changes, it takes a third of a second.

This has an important impact on new deploys. For example, a deploy of 32
VMs with the latest CPI (v58) takes 1548 seconds, but now with these
changes the deploy only takes 808 seconds.

This also improves the robustness of the deploy process. For large
deploys (40+ VMs), users would see errors (typically timeout errors and
OpenSSL errors). With these changes, users should see fewer such errors.

The CPI didn't always take so long to do deploys. Prior to the
introduction of the NSX-T Policy API, a 32 VM deploy would only take 837
seconds.

The slowdown is due to a vast increase in the number of libraries loaded
at runtime. The NSX-T Policy API introduced a large number of additional
libraries, and CPI calls spend much of their time loading libraries
(i.e. Ruby `require`).

This fix works by introducing a slimmed-down class that returns the
hard-coded string. That's right: the `info` vSphere CPI RPC call returns
the same string every time. By having a slimmed-down class, we bypass
the loading of libraries.

[#178336193] CPI calls like info should not load lots of files into memory

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>