🐞 NSX-T: Use LB Pool `id`, not `display_name`
When working with the NSX-T _Policy_ API, we previously used the pools'
`display_name` in the NSX-T URL paths. This was wrong; we need to use
the `id`.

This commit fixes that behavior. We also renamed variables (`_name` →
`_id`) to more accurately reflect what is being referred to.

Drive-by: we corrected English usage for some test descriptions

Fixes the following error:
```
The requested object : /infra/lb-pools/BOSH-CPI-test-8a84db51-90c5-46c2-b4cb-adf5de6cd4dd could not be found. Object identifiers are case sensitive.
```

[#178649776](https://www.pivotaltracker.com/story/show/178649776)

Co-authored-by: Gary Liu <garyliu@vmware.com>
Co-authored-by: Brian Cunnie <bcunnie@vmware.com>