🐞 Swagger produces correct NSX-T configuration
When running `bundle exec rake swagger...`, swagger
would use the same mustache configuration template for both Manager &
Policy APIs; however, this was a mistake. The configurations differ:
both `@base_path` and `base_url()` are different.

This commit remedies that mistake by separating the mustache
configuration files.

Swagger correctly bumps `target_id` from 64 bytes → 255 bytes for a
subset of calls.  `target_id` is one of the fields returned on the NSX-T
Policy API [Get groups for which the given VM is a member
](https://vdc-download.vmware.com/vmwb-repository/dcr-public/d6de7a5e-636f-4677-8dbd-6f4ba91fa5e0/36b4881c-41cd-4c46-81d1-b2ca3a6c693b/api_includes/method_GetGroupsForVM.html).

Our experience in our environment is that in the JSON payload returned
by the call `get_groups_for_vm()`, `target_id` is identical to the
`path` returned, and it's frequently longer than 64 bytes, which means
it would frequently error.

Fixes, when `get_groups_for_vm()` is called:
```
invalid value for "target_id", the character length must be smaller than or equal to 64
```

All future changes to dynamically generate code should be made via
swagger mustache templates so that we don't override important changes

[#178506726](https://www.pivotaltracker.com/story/show/178506726)

Signed-off-by: Julian Hjortshoj <jhortshoj@vmware.com>
Signed-off-by: Brian Cunnie <bcunnie@vmware.com>