Explictly test LB Server Pools without ports
Ports are optional in the NSX-T Load Balancer server pool specifications
(via VM extensions). Previously we had wrongly indicated they were required.
However, it's important to allow ports to be skipped in cases like the
Cloud Foundry TCP Router, which has many ports, and the NSX-T API only
allows the specification of one port.

When the port is skipped, the port is set to nil, and the NSX-T Virtual
Server backing the Load Balancer will forward the traffic to the backing
instance on the same port that the Virtual Server received the traffic
(e.g. 80, 443, 2222).

The code has always handled the absence of ports correctly, we are just
adding tests to draw visibility to this feature for future developers.

[#178981294](https://www.pivotaltracker.com/story/show/178981294)

Co-authored-by: Brian Cunnie <bcunnie@vmware.com>
Co-authored-by: Mark Stokan <stokanm@vmware.com>