Reorder network inspection code
- New logic tests each network to see if it matches all criteria instead
of gathering a collection of networks and then progressively narrowing
that collection by criteria.
- We believe that this may help to address
[PKS-877](https://pivotal-spike.atlassian.net/browse/PKS-877) in which
churn testing causes the CPI code to spend excessive time in exponential
backoff looking for networks that have been deleted. It should also
reduce the total number of API calls in cases where there are a lot of
networks.

[#178814962](https://www.pivotaltracker.com/story/show/178814962)

Co-authored-by: Julian Hjortshoj <hjortshojj@vmware.com>
Co-authored-by: Mark Stokan <mstokan@pivotal.io>