Remove obsolete CD-ROM remounting code
- Previous CPI versions modified `env.json` and re-attached `env.iso` to
pass along disk metadata to the BOSH agent during `attach_disk` calls.
- After updating the vSphere CPI to use v2 of the interface, we now pass
disk metadata through the BOSH director by returning it from the
`attach_disk` call, so re-attaching the CD-ROM is no longer necessary.
- This commit removes the obsolete code.
- Note that the CD-ROM mounting code is still required at VM creation
time so that the BOSH agent can learn the IP address of the VM.
- Note also that `env.json` is still written out to the datastore even
though we no longer need it. We decided to leave it in case customer use
it for debugging or script automation.

[#179679846](https://www.pivotaltracker.com/story/show/179679846)