Update find_disk to use the `persistent_datastore_cluster_pattern` if specified in cpi config
We were only searching using `persistent_pattern` before. This did not cause us to NOT find the
disk, because we would then search all available datastores. It would cause a failure
if you failed to specify the `persistent_pattern` because we would fail when trying
to use `nil` as a Regular Expression.

[#180286111]

[Fixes #306]

Signed-off-by: Joseph Palermo <jpalermo@pivotal.io>