Search for just the network we need
The existing code has a big problem: It knows the name of the network
that it needs to find, and can ask the vCenter for that specific
network, but chooses to get _all_ networks, and then do client-side
filtering. This has a time-of-check/time-of-use problem in that we
iterate over all of the networks provided to compare their name with the
name of the network we need to operate on.

This means that we must fetch the properties for each network as we go
(the initial call doesn't fetch them for us). If any of the networks in
the list have been deleted since we got the list, we retry for tens of
seconds to ensure that the network isn't going to appear because it was
scheduled to be created and hasn't appeared yet.

This commit asks for only the network that we're interested in, so that
we reduce the amount of time we waste wading through networks that we
don't care about while searching for the network that we're interested
in.

NOTE: If the network is "in a container", that is, has `/` characters in
its name, we still use the old behavior. We should probably find out how
to and then fix that.

[#178814962] PKS #877 speedup vSphere (NSX-T) network fetching logic

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>
Signed-off-by: Brian Upton <bupton@vmware.com>