Do not retry RetrieveProperties for OpaqueNetwork
We reverted the previous attempt to fix this because we think it
may fail to correctly find networks that are nested.

We think this fix will prevent retries when trying to get properties
for networks that have been deleted. This change is scoped very
narrowly to certain types of NSX-T port groups, but it is possible
we may need to expand this to objects other than OpaqueNetworks
in the future.

It is possible that we fail to retry now when in fact we got an
incorrect "Network Not Found" error, but we have no data on how
often that happens.

[#178814962] PKS #877 speedup vSphere (NSX-T) network fetching logic

Signed-off-by: Joseph Palermo <jpalermo@pivotal.io>