🐞Fix: the CPI is removing server pool members (VMs) from unmanaged pools
Problem: Any VM that shares the same private IP as a bosh managed VM and
is also a member of a server pool will be removed from the pool by the
CPI. The CPI is iterating over all pools and not limited to just pools
provided in the configuration when recreating or deleting an instance.

Fix:
When adding a VM to a server pool we set the display_name on the Pool
Member to the VM CID of the VM being added.
When deleting a VM, we only delete server pool members who have a
display name equals to the VM CID. The field added on the previous
commit "cpi_metadata_version" will make this change backwards
compatible: the VMs created before this change will be removed from all
the server pools (cpi_metadata_version = 0), the VMs created after this change will be removed
only from bosh managed server pools (cpi_metadata_version > 0).

[#180286132] (https://www.pivotaltracker.com/story/show/180286132)