Dockerfile: pull ruby-install from web, not submodule
We are unsure of why `ruby-install`, unlike every other package
installed in the Dockerfile, was pulled from a local submodule. The
reasons are lost in the sands of time.

Curiously, there is only one submodule in this repo, and that's for
`ruby-install`. We should remove that submodule in the future.

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>