Docker image `bosh/vsphere-vcpi` uses vanilla nvim
Previously we introduced Luan's customized neovim configuration, which
had a mixed reception. This commit removes the customized configuration.

I speculate that the long neovim start-up time while it downloaded
plugins was the main source of disappointment. The vanilla neovim
configuration does not download plugins, and is ready to go right away.

Drive-bys:

I removed the git submodule for `ruby-install`, used solely when
creating the `bosh/vsphere-vcpi` Docker image. An earlier commit
deprecated the use of the submodule, whose sole purpose appears to have
been to pin `ruby-install` to a known working version. We now build the
latest version of `ruby-install`, which appears to work fine.