🐞 Eliminate NSX-T Race Condition
There's a race condition: when the NSX-T Policy API Migration Mode is
enabled, and a VM is added to a server pool via the Policy API and then
subsequently added to server pools via the Management API, then an
error, "There are more than one pool member which have the same IP
address" may occur.

This commit resolves the race condition by re-ordering the operations so
that the VM is added to the pool via the Management API first, the
Policy API, second.

The error is apt to occur when there's a delay between adding the VM to
the pool via the Policy API and the Management API, for example, when
running the tests from a workstation with slow connection or when
debugging a failed run using `pry-byebug`.

When we manually inserted a 60 second sleep between the operations, the
failure occurred 10/10 times. When we removed the sleep, it succeeded
4/4 times.

Fixes:
```
NSXT::ApiCallError:
       [{"httpStatus"=>"BAD_REQUEST", "error_code"=>23613, "module_name"=>"LOAD-BALANCER", "error_message"=>"There are more than one pool member which have the same IP address 169.254.178.190 and port 80."}]
```

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>