Generate certificates compliant w/ newer TLS reqs
When trying to upgrade to newer versions of Ubuntu (which include newer
versions of OpenSSL), we were encountering a failure where the private
key of the certificate used with the HTTP server in tests was too short.

> [2022-04-13 19:09:23] ERROR OpenSSL::SSL::SSLError: SSL_CTX_use_certificate: ee key too small
>  /usr/local/lib/ruby/2.6.0/webrick/server.rb:259:in `initialize'

Rather than hardcode new values, we wrote a somewhat-short method to
generate certificates on the fly.

Now, we have just a signing CA which signs the leaf cert, and an
unrelated CA which is not part of the signing chain. This makes the
certificate(:failure) case more clear in that it doesn't include the
signing CA needed to trust the leaf cert.

[#181892208] vSphere CPI should use more secure certificates in testing (so we can upgrade Docker image's Ubuntu)

Signed-off-by: Brian Upton <bupton@vmware.com>