Only one source of truth for project Ruby version
Prior to this commit, the Ruby version was specified in a number of
different locations in the project. Bumping the Ruby version was an
error-prone task.

This commit fixes that by installing one source of truth:
`.ruby-version` at the project root. We dispense with
`PROJECT_RUBY_VERSION` set via `.envrc`, and checking that in numerous
places.

Drive-by: we removed the `scripts/test_*_coverage` scripts; they weren't
called anywhere, and their purpose is easily fulfilled by setting
`COVERAGE=true` when running `rspec` (which we noted in a comment).

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>