Use environment variables for bundle deployment
Newer versions of bundler are not looking at the right vendor directory
when the CPI is packaged or run using the `--deployment` flag. Using the
environment variables instead should make it look at where we have
always vendored our gems.

See this commit in the Azure CPI for more details:
https://github.com/cloudfoundry/bosh-azure-cpi-release/commit/ffde6ed7cf3e550a9e5ee1c65e9bcf4eb5fd4adf

Signed-off-by: Brian Upton <bupton@vmware.com>