Place persistent disks properly with vertical storage
Vertical storage is when Datastores are not universally accessible from
all cluster within a Datacenter.

This poses a problem for users attempting to deploy to a multi-Cluster
Datacenter: until now, the CPI assumed that a datastore was universally
accessible, and did not make sure that when it created a disk, it
created the disk on a datastore that was accessible to the VM where it
was to be attached.

This commit fixes that problem: before creating a persistent disk, it
checks to make sure that the VM, or, more specifically, the Cluster in
which the VM resides, has access to the datastore on which the disk is
created.

fixes:
```
CPI 'attach_disk' method responded with error: CmdError{"type":"Unknown","message":"Invalid configuration for device '0'.","ok_to_retry":false}
```

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>