update vsphere cpi to deploy a jammy stemcell
the create-env fails because we're using a really old xenial stemcell
but bosh deployment repo switched to jammy as a default. that switch
also updated compiled releases to be jammy based and currently that
breaks the create-env.

update vsphere cpi to test on jammy as that is what our target is going
forward.