Revert "Increase ram to fix tests"
This commit should not be necessary. This failure scenario only occurs
if the first test in the multi_cluster_overflow spec fails. If we want
to prevent this failure scenario we can make that test slightly smarter
(save off the VM in case it is created and ensure it is deleted).

This reverts commit 2e0514712f60ad103c1cebb2f6435861c9f887f2.