Remove policy API requirement from migration prop
The policy_api_migration_mode property as currently written does not
require that the use_policy_api property be set to true. In general, it
is intended to be used only when use_policy_api is set to false.
However, in practice it does not matter whether it is enabled or not
because the policy_api_migration_mode property takes precedence.