Better performance: trim NSX auto-generated objects
BOSH vSphere CPI is slower when NSX is configured, increasing deploy
times by as much as 10%. Specifically, the CPI call
[`set_vm_metadata`](https://bosh.io/docs//cpi-api-v2-method/set-vm-metadata/)
can double the amount of time to complete (e.g. from 5 seconds to 10
seconds).

We can mitigate this by only using the APIs we need, for earlier
investigation showed that much of the additional time was spent loading
the thousands of NSX classes.

This commit makes the following reductions:

- NSX Policy
  - Endpoints (e.g. `/infra/lb-pools`): 1299 → 8
  - Models/Ojbects (e.g. `SegmentPort`): 1344 → 75
- NSX Management (deprecated)
  - Endpoints (e.g. `/infra/lb-pools`): 872 → 24
  - Models/Ojbects (e.g. `SegmentPort`): 1260 → 146

We reduced the total number of non-test Ruby files in the vSphere CPI
3049 → 348, which represents an 88% reduction in the number of files.

Notes on endpoint changes:

- `read_lb_pool_0()` → `read_lb_pool()` because the original
  `read_lb_pool`, generated from the endpoint
  `/global-infra/lb-pools/{lb-pool-id}`, is no longer instantiated. We
  don't use that endpoint.
- Similarly with `update_lb_pool_0()`.
- Similarly, `get_tier1_segment_port_0()` → `get_tier1_segment_port` due
  to unused endpoint
  `/global-infra/tier-1s/{tier-1-id}/segments/{segment-id}/ports/{port-id}`
- And `patch_tier1_segment_port_0`, too.

Signed-off-by: Brian Cunnie <bcunnie@vmware.com>