add test to vm selection pipeline for new fallback_disk_placements attr
the review for the PR pointed out that we're missing coverage here. good
thing it did because writing the test uncovered an oversight.

The problem was repeated calls to `pipeline.each.first`. This is not
idempotent and can return varying ordering on subsequent calls. So
dropping the first item is not  necessarily what we want to do. What we
want to do is to delete `result` from the returned array and use the rest
to avoid that `result` itself is part of the list of fallbacks.

```
    127:         logger.debug("Found #{result.inspect} for #{disk.inspect}")
    128:         vm_placement.disk_placement = result unless disk.existing_datastore_name
    129:         vm_placement.fallback_disk_placements =  pipeline.each.to_a - [result] if disk.ephemeral?
 => 130:         require 'pry'; binding.pry
    131:
    132:         logger.debug("Found alternative disk placements: #{vm_placement.fallback_disk_placements}") if disk.ephemeral?
    133:         result.free_space -= disk.size
    134:         vm_placement.balance_score_set << result
    135:         vm_placement.migration_size += disk.size if disk.existing_datastore_name

[1] pry(VSphereCloud::VmPlacementSelectionPipeline)> pipeline.each.first
I, [2023-03-10T11:06:23.157686 #585753]  INFO -- : Initiating Disk Allocator with Storage Criteria: Space Req: 10 Target Pattern: .*
        Existing-DS Pattern:
=> <Datastore: / fake-ds-cl1-2>
[2] pry(VSphereCloud::VmPlacementSelectionPipeline)> pipeline.each.first
I, [2023-03-10T11:06:24.441858 #585753]  INFO -- : Initiating Disk Allocator with Storage Criteria: Space Req: 10 Target Pattern: .*
        Existing-DS Pattern:
=> <Datastore: / fake-ds-cl1-2>
[3] pry(VSphereCloud::VmPlacementSelectionPipeline)> pipeline.each.first
I, [2023-03-10T11:06:25.040907 #585753]  INFO -- : Initiating Disk Allocator with Storage Criteria: Space Req: 10 Target Pattern: .*
        Existing-DS Pattern:
=> <Datastore: / fake-ds-cl1-1>
```