Create VM: try second cluster when first fails
If, during a VM creation, we have several clusters to choose from when
we place the VM, and if placing the VM in the first cluster fails, then
we never test the second or follow-on clusters.

This commit fixes that. We now attempt to place the VM in the subsequent
clusters.

This is one of those elegant commits that fixes a problem by simplifying
the code (3 lines deleted, 1 line modified).

Fixes, during `be rspec spec/integration`:
```
rspec ./spec/integration/multi_cluster_overflow_spec.rb:46 # given 2 clusters with a datastore and a resource pool when no cluster has enough reservable memory should raise error for insufficient available Memory
rspec ./spec/integration/multi_cluster_overflow_spec.rb:71 # given 2 clusters with a datastore and a resource pool when the first cluster has enough reservable memory should create a VM on the first cluster
```

Signed-off-by: Brian Upton <bupton@vmware.com>
Signed-off-by: Brian Cunnie <bcunnie@vmware.com>