CI: make testbed clean-up less brittle
As part of our continuous integration (CI), we create testbeds (vCenters,
ESXi hosts, NSX Managers), that we later clean-up (destroy) so that we
don't exceed our quota; however, our Concourse task that destroys the
testbeds is brittle, and fails for unimportant reasons.

This commit addresses that by turning off `-e` (bash's `errexit`) for
the fragile command (`nimbus-ctl --testbed list`). This command is quick
to throw an error, but it's usually not important enough to derail the
task (e.g. "Command 'list' does not finish within 600 seconds on
sc2-04-vc17").