Remove spec checking for exact Bundler version
This spec was introduced because if the version of Bundler in
Gemfile.lock becomes out-of-sync with the version of Bundler provided by
the Ruby package, Bundler would try to install the specified version and
thus break in airgapped environment.

Since that time, we've introduced a change in the packaging script for
the CPI that should ensure the Gemfile.lock version of Bundler is always
used. See this commit: e5d323a868dd8b597aca24b084b480a58edafb80